Replacing unique docker client by a docker client factory.
From an error that we were getting, I suspect that the client might not be thread safe.

Just in case, for the moment we will generate one client per thread using the factory.
If in the future we want to come back to the shared client (the current version might be less memory efficient), this can be done easily slightly modifying this factory.